+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### [0.4.0] - 2025-10-31 - Arithmetic Expansions
 
 - Added arithmetic functions:
+  - `productOfDigits(n)` – Computes the product of the digits of an integer n
   - `sumOfDigits(n)` – Computes the sum of the digits of an integer n
 
 ## [0.3.0] - 2025-10-31 - Prime and Function Expansions